refactor: styled props 넘기는 방식 수정 (console 에러)
