+++ src/components/store-info/index.tsx
